Refactored index.js to separate out app.js for ease of testing
